Moved to ASP.NET
Removed Firebase.  Static website will be deleted later.  Digital Ocean now has a spa/webapp service that has a free tier and there pro tier has a great price vs the other.  Was going to go with firebase but want i really wanted to do with this program is show off my full stack skills so ASP.NET is the better thing to be working on.